==> 01-Plan_your_trip_with_Kayak.ipynb <==
import pandas as pd 
import numpy as np
import requests 
import json
from tqdm import tqdm
from bs4 import BeautifulSoup
import scrapy
from scrapy.crawler import CrawlerProcess
import plotly.express as px
import logging
import os 

top_35_cities=["Mont Saint Michel", "St Malo", "Bayeux", "Le Havre", "Rouen", "Paris", "Amiens", "Lille", "Strasbourg",
"Chateau du Haut Koenigsbourg","Colmar", "Eguisheim", "Besancon", "Dijon","Annecy", "Grenoble", "Lyon", "Gorges du Verdon",
"Bormes les Mimosas", "Cassis", "Marseille", "Aix en Provence", "Avignon", "Uzes", "Nimes", "Aigues Mortes",
"Saintes Maries de la mer", "Collioure", "Carcassonne", "Ariege", "Toulouse", "Montauban", "Biarritz", "Bayonne",
"La Rochelle"]

df= pd.DataFrame()

for city in tqdm(top_35_cities):
    url="https://nominatim.openstreetmap.org/search?q="+city+"&format=json&countrycodes=fr&limit=1"
    result=requests.get(url)
    dict=result.json()[0]
    dict["city"]=city
    df=df.append(dict,ignore_index=True)

df.dtypes

df.head()

df2= pd.DataFrame()

for index, row in tqdm(df.iterrows()):
    lat=row["lat"]
    lon=row["lon"]
    city=row["city"]
    url="https://api.openweathermap.org/data/2.5/onecall?lat="+lat+"&lon="+lon+"&exclude=current,minutely,hourly&units=metric&appid="
    result=requests.get(url)
    dict_=result.json()
    df3= pd.DataFrame()
    
    for i in range(0,8):
        
        number=str(i)
        
        t_min=dict_["daily"][i]["temp"]["min"]
        t_max=dict_["daily"][i]["temp"]["max"]
        t_day=dict_["daily"][i]["temp"]["day"]
        t_eve=dict_["daily"][i]["temp"]["eve"]
        t_night=dict_["daily"][i]["temp"]["night"]
        pressure=dict_["daily"][i]["pressure"]
        humid=dict_["daily"][i]["humidity"]
        dew_point=dict_["daily"][i]["dew_point"]
        wind_speed=dict_["daily"][i]["wind_speed"]
        wind_deg=dict_["daily"][i]["wind_deg"]
        wind_gust=dict_["daily"][i]["wind_gust"]
        clouds=dict_["daily"][i]["clouds"]
        
        keys=["t_min_"+number,"t_max_"+number,"t_day_"+number,"t_eve_"+number,"t_night_"+number,
              "pressure_"+number,"humid_"+number,"dew_point_"+number,"wind_speed_"+number,"wind_deg_"+number,
              "wind_gust_"+number,"clouds"+number]
        values=[t_min,t_max,t_day,t_eve,t_night,pressure,humid,dew_point,wind_speed,wind_deg,wind_gust,clouds]
        
        #dictionary=dict(zip(keys, values))
        #dictionary["city"]=city
        df_f = pd.DataFrame(values,index =keys)
        frames = [df3, df_f]
        df3=pd.concat(frames)
    
    df3=df3.T
    df2=df2.append(df3, ignore_index=True)
        

df2.head()
        

final_df=pd.concat([df, df2], axis=1)
final_df.head()

final_df['T_Average'] = final_df[['t_min_0', 't_max_0', 't_day_0', 't_eve_0', 't_night_0',
                            't_min_1', 't_max_1', 't_day_1', 't_eve_1', 't_night_1',
                           't_min_2', 't_max_2', 't_day_2', 't_eve_2', 't_night_2',
                            't_min_3', 't_max_3', 't_day_3', 't_eve_3', 't_night_3',
                            't_min_4', 't_max_4', 't_day_4', 't_eve_4', 't_night_4',
                            't_min_5', 't_max_5', 't_day_5', 't_eve_5', 't_night_5',
                            't_min_6', 't_max_6', 't_day_6', 't_eve_6', 't_night_6',
                            't_min_7', 't_max_7', 't_day_7', 't_eve_7', 't_night_7',                           
                           ]].mean(axis=1)

final_df['T_Average'] = final_df['T_Average'].astype(float)
final_df['lat'] = final_df['lat'].astype(float)
final_df['lon'] = final_df['lon'].astype(float)

final_df=final_df.sort_values(by='T_Average', ascending=False).reset_index(drop=True)
final_df.head()

final_df.to_excel("cities.xlsx")
final_df.to_csv("cities.csv")

fig = px.scatter_mapbox(final_df, lat="lat", lon="lon", color="T_Average", zoom =4,size="T_Average",size_max=10,
                        mapbox_style="open-street-map", color_continuous_scale = 'Bluered', range_color = [8,20.0],
                       )
fig.show()
