==> 04-Plan_your_trip_with_Kayak.ipynb <==
# Project: planning my next holidays ☀️

Let's create a script that allows to get some information about all the hotels in a given city on <a href="https://www.booking.com" target="_blank">www.booking.com</a> 🧙

**We strongly recommend that you use Scrapy, it will be much easier!**

You can scrap as many information as you want, but we suggest that you get at least:

* The hotel name, 
* The url to its booking.com page, 
* Its coordinates: latitude and longitude,
* The score given by the website users,
* The text description of the hotel.

Then, you can execute this script for several cities from yesterday's list. Make sure you save the results in different files for each city and that the name of the city is stored in the filename (for later purposes 😉).

import pandas as pd 
from sqlalchemy import create_engine
import plotly.express as px
import logging
import os 

engine = create_engine("postgresql+psycopg2://", echo=True)

path_to_file="/Users/saas1/OneDrive/Escritorio/introduction_to_ python_for_data_science/Projects/Project_3/Project/"

data_frame_1=pd.read_csv (path_to_file +"/cities.csv",index_col='Unnamed: 0')
data_frame_2=pd.read_excel(path_to_file +"/hotel.xlsx",index_col='Unnamed: 0')

data_frame_2.shape


data_frame_2= data_frame_2.drop(['url', 'description'], axis=1)
data_frame_2.dropna(subset=['score'], inplace=True)
data_frame_2['score'] = data_frame_2['score'].str.replace(',','.')
data_frame_2['score'] = data_frame_2['score'].astype(float)
data_frame_2.head()

data_frame_1.to_sql("cities",engine)


best_5=["Collioure", "Marseille", "Cassis","Bayonne", "Bormes les Mimosas"]

data_frame_2=data_frame_2[data_frame_2['city'].isin(best_5)]


data_frame_2.to_sql("hotels_france",engine)

from sqlalchemy.sql import text

stmt = text("SELECT * FROM hotels_france ")

df_hotels = pd.read_sql(
        stmt,
        engine
    )



from sqlalchemy.sql import text

stmt = text("SELECT * FROM cities ")

df_cities = pd.read_sql(
        stmt,
        engine
    )


df_hotels=df_hotels.sort_values(by='score', ascending=False).reset_index(drop=True)

best_5=["Collioure", "Marseille", "Cassis","Bayonne", "Bormes les Mimosas"]
hotels_df=pd.DataFrame()

for i in best_5:
    hotels_df=hotels_df.append(df_hotels[df_hotels["city"]==i].iloc[0:20,:])

hotels_df.reset_index(drop=True)

fig = px.scatter_mapbox(hotels_df, title="Best Hotels",lat="lat", lon="lon", color="score", zoom =5.5,size="score",size_max=10,
                        mapbox_style="open-street-map", color_continuous_scale = 'Bluered', range_color = [8,10.0],hover_name="name"
                       )
fig.show()


df_cities=df_cities.iloc[0:5,]

fig = px.scatter_mapbox(df_cities, lat="lat", lon="lon", color="T_Average", zoom =5.5,size="T_Average",size_max=10,
                        mapbox_style="open-street-map", color_continuous_scale = 'OrRd', range_color = [8,20.0],
                       title="Best Cities",hover_name="city")
fig.show()
